Moved tests.util to test_helpers
due to tests/__init__ and old Django 1.5
